+++ CHANGES.rst
@@ -11,6 +11,9 @@ Changelog
 - Export and import complete sites or content trees with configurable types, depth and path (#40).
   [pbauer]
 
+- Added option to export blobs as blob paths (#50).
+  [pbauer, maurits]
+
 - Fixed creating missing folder structure (#45).
   [maurits]